Refactored operations on data for graph into graph_data.py file
