+++ src/value_type.py
@@ -1,0 +1,30 @@
+from enum import Enum
+
+class ValueType(Enum):
+    INTEGER = (11, "Integer", [1])
+    CAT_SINGLE = (21, "Categorical (single)", [3, 4])
+    CAT_MULT = (22, "Categorical (multiple)", [3, 4])
+    CONT = (31, "Continuous", [1])
+    TEXT = (41, "Text", [])
+    DATE = (51, "Date", [])
+    TIME = (61, "Time", [])
+    COMPOUND = (101, "Compound", [])
+
+    def __init__(self, type_id, label, supported_graphs):
+        self.type_id = type_id
+        self.label = label
+        self.supported_graphs = supported_graphs
+
+    def __new__(cls, *values):
+        obj = object.__new__(cls)
+        # first value is canonical value
+        obj._value_ = values[0]
+        obj._all_values = values
+        return obj
+
+    def __repr__(self):
+        return "<%s.%s: %s>" % (
+            self.__class__.__name__,
+            self._name_,
+            ", ".join([repr(v) for v in self._all_values]),
+        )
